move swing related stuff out of runtime package into org.antlr.v4.gui
