+++ runtime/Java/pom.xml
@@ -11,14 +11,6 @@
   <name>ANTLR 4 Runtime</name>
   <description>The ANTLR 4 Runtime</description>
 
-  <dependencies>
-    <dependency>
-      <groupId>org.abego.treelayout</groupId>
-      <artifactId>org.abego.treelayout.core</artifactId>
-      <version>1.0.1</version>
-    </dependency>
-  </dependencies>
-
   <build>
     <sourceDirectory>src</sourceDirectory>
     <plugins>
@@ -38,7 +30,7 @@
           </execution>
         </executions>
       </plugin>
-      <plugin>   
+      <plugin>
         <groupId>org.apache.felix</groupId>
         <artifactId>maven-bundle-plugin</artifactId>
         <version>2.5.4</version>
@@ -46,7 +38,7 @@
           <execution>
             <id>bundle-manifest</id>
             <phase>process-classes</phase>
-            <configuration>            
+            <configuration>
               <instructions>
                 <Bundle-SymbolicName>org.antlr.antlr4-runtime-osgi</Bundle-SymbolicName>
                 <Implementation-Title>ANTLR 4 Runtime</Implementation-Title>
@@ -55,19 +47,19 @@
                 <Implementation-Version>${project.version}</Implementation-Version>
               </instructions>
             </configuration>
-            <goals>    
+            <goals>
               <goal>manifest</goal>
-            </goals>   
+            </goals>
           </execution>
         </executions>
-      </plugin>            
+      </plugin>
       <plugin>
         <artifactId>maven-jar-plugin</artifactId>
         <version>2.4</version>
         <configuration>
-          <archive>  
+          <archive>
             <manifestFile>${project.build.outputDirectory}/META-INF/MANIFEST.MF</manifestFile>
-          </archive> 
+          </archive>
         </configuration>
       </plugin>
       <plugin>
@@ -78,14 +70,14 @@
           <execution>
             <phase>package</phase>
             <configuration>
-              <dependencyReducedPomLocation>${project.build.directory}/dependency-reduced-pom.xml</dependencyReducedPomLocation>        
-            </configuration>            
+              <dependencyReducedPomLocation>${project.build.directory}/dependency-reduced-pom.xml</dependencyReducedPomLocation>
+            </configuration>
             <goals>
               <goal>shade</goal>
             </goals>
           </execution>
         </executions>
-      </plugin>      
+      </plugin>
     </plugins>
   </build>
 </project>
