+++ runtime/Java/src/org/antlr/v4/runtime/tree/Trees.java
@@ -37,12 +37,7 @@ import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.misc.Interval;
 import org.antlr.v4.runtime.misc.Predicate;
 import org.antlr.v4.runtime.misc.Utils;
-import org.antlr.v4.runtime.tree.gui.TreePostScriptGenerator;
-import org.antlr.v4.runtime.tree.gui.TreeTextProvider;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -51,41 +46,6 @@ import java.util.List;
 
 /** A set of utility routines useful for all kinds of ANTLR trees. */
 public class Trees {
-
-	public static String getPS(Tree t, List<String> ruleNames,
-							   String fontName, int fontSize)
-	{
-		TreePostScriptGenerator psgen =
-			new TreePostScriptGenerator(ruleNames, t, fontName, fontSize);
-		return psgen.getPS();
-	}
-
-	public static String getPS(Tree t, List<String> ruleNames) {
-		return getPS(t, ruleNames, "Helvetica", 11);
-	}
-
-	public static void writePS(Tree t, List<String> ruleNames,
-							   String fileName,
-							   String fontName, int fontSize)
-		throws IOException
-	{
-		String ps = getPS(t, ruleNames, fontName, fontSize);
-		FileWriter f = new FileWriter(fileName);
-		BufferedWriter bw = new BufferedWriter(f);
-		try {
-			bw.write(ps);
-		}
-		finally {
-			bw.close();
-		}
-	}
-
-	public static void writePS(Tree t, List<String> ruleNames, String fileName)
-		throws IOException
-	{
-		writePS(t, ruleNames, fileName, "Helvetica", 11);
-	}
-
 	/** Print out a whole tree in LISP form. {@link #getNodeText} is used on the
 	 *  node payloads to get the text for the nodes.  Detect
 	 *  parse trees and extract data appropriately.
@@ -104,38 +64,23 @@ public class Trees {
 		return toStringTree(t, ruleNamesList);
 	}
 
-	/** Print out a whole tree in LISP form. Arg nodeTextProvider is used on the
-	 *  node payloads to get the text for the nodes.
-	 *
-	 *  @since 4.5.1
-	 */
-	public static String toStringTree(Tree t, TreeTextProvider nodeTextProvider) {
-		if ( t==null ) return "null";
-		String s = Utils.escapeWhitespace(nodeTextProvider.getText(t), false);
-		if ( t.getChildCount()==0 ) return s;
-		StringBuilder buf = new StringBuilder();
-		buf.append("(");
-		s = Utils.escapeWhitespace(nodeTextProvider.getText(t), false);
-		buf.append(s);
-		buf.append(' ');
-		for (int i = 0; i<t.getChildCount(); i++) {
-			if ( i>0 ) buf.append(' ');
-			buf.append(toStringTree(t.getChild(i), nodeTextProvider));
-		}
-		buf.append(")");
-		return buf.toString();
-	}
-
 	/** Print out a whole tree in LISP form. {@link #getNodeText} is used on the
 	 *  node payloads to get the text for the nodes.
 	 */
 	public static String toStringTree(final Tree t, final List<String> ruleNames) {
-		return toStringTree(t, new TreeTextProvider() {
-			@Override
-			public String getText(Tree node) {
-				return getNodeText(node, ruleNames);
-			}
-		});
+		String s = Utils.escapeWhitespace(getNodeText(t, ruleNames), false);
+		if ( t.getChildCount()==0 ) return s;
+		StringBuilder buf = new StringBuilder();
+		buf.append("(");
+		s = Utils.escapeWhitespace(getNodeText(t, ruleNames), false);
+		buf.append(s);
+		buf.append(' ');
+		for (int i = 0; i<t.getChildCount(); i++) {
+			if ( i>0 ) buf.append(' ');
+			buf.append(toStringTree(t.getChild(i), ruleNames));
+		}
+		buf.append(")");
+		return buf.toString();
 	}
 
 	public static String getNodeText(Tree t, Parser recog) {
@@ -204,7 +149,7 @@ public class Trees {
 		if ( t==null || u==null || t.getParent()==null ) return false;
 		Tree p = u.getParent();
 		while ( p!=null ) {
-			if ( t == p ) return true;
+			if ( t==p ) return true;
 			p = p.getParent();
 		}
 		return false;
