+++ tool/src/org/antlr/v4/gui/JFileChooserConfirmOverwrite.java
@@ -27,7 +27,7 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  *  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.antlr.v4.runtime.misc;
+package org.antlr.v4.gui;
 
 import javax.swing.*;
 import java.io.File;
